add simplified QuestionsListView (#28)
